init foundry-wallets
