+++ crates/cast/Cargo.toml
@@ -24,6 +24,7 @@ foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
 foundry-evm.workspace = true
+foundry-wallets.workspace = true
 
 alloy-dyn-abi.workspace = true
 alloy-json-abi.workspace = true
@@ -78,8 +79,8 @@ criterion = "0.5"
 
 [features]
 default = ["rustls"]
-rustls = ["foundry-cli/rustls"]
-openssl = ["foundry-cli/openssl"]
+rustls = ["foundry-cli/rustls", "foundry-wallets/rustls"]
+openssl = ["foundry-cli/openssl", "foundry-wallets/openssl"]
 asm-keccak = ["alloy-primitives/asm-keccak"]
 
 [[bench]]
