+++ crates/cast/bin/cmd/wallet/mod.rs
@@ -9,12 +9,12 @@ use ethers_signers::{
     LocalWallet, MnemonicBuilder, Signer,
 };
 use eyre::{Context, Result};
-use foundry_cli::opts::{RawWallet, Wallet};
 use foundry_common::{
     fs,
     types::{ToAlloy, ToEthers},
 };
 use foundry_config::Config;
+use foundry_wallets::{RawWallet, Wallet};
 use serde_json::json;
 use std::path::Path;
 use yansi::Paint;
