+++ crates/cli/Cargo.toml
@@ -15,6 +15,7 @@ foundry-common.workspace = true
 foundry-config.workspace = true
 foundry-debugger.workspace = true
 foundry-evm.workspace = true
+foundry-wallets.workspace = true
 
 foundry-compilers = { workspace = true, features = ["full"] }
 
@@ -24,26 +25,17 @@ alloy-primitives.workspace = true
 
 ethers-core.workspace = true
 ethers-providers.workspace = true
-ethers-signers = { workspace = true, features = ["aws", "ledger", "trezor"] }
 
-rusoto_core = { version = "0.48", default-features = false }
-rusoto_kms = { version = "0.48", default-features = false }
-
-async-trait = "0.1"
 clap = { version = "4", features = ["derive", "env", "unicode", "wrap_help"] }
 color-eyre.workspace = true
 dotenvy = "0.15"
 eyre.workspace = true
-hex = { workspace = true, features = ["serde"] }
 indicatif = "0.17"
-itertools.workspace = true
 once_cell = "1"
 regex = { version = "1", default-features = false }
-rpassword = "7"
 serde.workspace = true
 strsim = "0.10"
 strum = { workspace = true, features = ["derive"] }
-thiserror = "1"
 tokio = { version = "1", features = ["macros"] }
 tracing-error = "0.2"
 tracing-subscriber = { workspace = true, features = ["registry", "env-filter", "fmt"] }
@@ -55,5 +47,5 @@ tempfile = "3.7"
 
 [features]
 default = ["rustls"]
-rustls = ["ethers-providers/rustls", "rusoto_core/rustls"]
-openssl = ["ethers-providers/openssl", "foundry-compilers/openssl"]
+rustls = ["ethers-providers/rustls", "foundry-wallets/rustls"]
+openssl = ["ethers-providers/openssl", "foundry-compilers/openssl", "foundry-wallets/openssl"]
