+++ crates/cli/src/opts/mod.rs
@@ -3,11 +3,9 @@ mod chain;
 mod dependency;
 mod ethereum;
 mod transaction;
-mod wallet;
 
 pub use build::*;
 pub use chain::*;
 pub use dependency::*;
 pub use ethereum::*;
 pub use transaction::*;
-pub use wallet::*;
