+++ crates/wallets/src/lib.rs
@@ -1,0 +1,11 @@
+#[macro_use]
+extern crate tracing;
+
+pub mod error;
+pub mod multi_wallet;
+pub mod raw_wallet;
+pub mod wallet;
+
+pub use multi_wallet::MultiWallet;
+pub use raw_wallet::RawWallet;
+pub use wallet::{Wallet, WalletSigner};
